+++ src_ts/components/pages/interventions/intervention-tab-pages/intervention-details/technical-guidance-capacity/technicalGuidance.models.ts
@@ -1,6 +1,5 @@
 import {ModelsBase} from '../../common/models/models.base';
-import {InterventionPermissionsFields, Intervention} from '../../common/models/intervention-types';
-
+import {InterventionPermissionsFields, Intervention} from '../../common/models/intervention.types';
 
 export class TechnicalDetails extends ModelsBase {
   constructor(intervention: Intervention) {
