+++ src_ts/components/pages/interventions/intervention-tab-pages/intervention-details/technical-guidance-capacity/technicalGuidance.selectors.ts
@@ -1,5 +1,5 @@
 import {createSelector} from 'reselect';
-import {Intervention, Permission, InterventionPermissionsFields} from '../../common/models/intervention-types';
+import {Intervention, Permission, InterventionPermissionsFields} from '../../common/models/intervention.types';
 import {TechnicalDetails, TechnicalDetailsPermissions} from './technicalGuidance.models';
 import {currentInterventionPermissions, currentIntervention} from '../../common/selectors';
